make Unity Dots features into a UPM package
