+++ UnitySDK/Packages/OpenBrushToolkitDots/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 323b8d051e809004bbbdf8b5d674ed24
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
