+++ UnitySDK/Packages/OpenBrushToolkitDots/Editor/Jobs/TriangleArrayCopyJob.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 43ff59cef122a2b499999973ea9e7428
+guid: 357741bf1569b48488ce0e13b49e77b8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
