+++ UnitySDK/Packages/OpenBrushToolkitDots/Editor/Jobs/TriangleCountJob.cs
