+++ UnitySDK/Packages/OpenBrushToolkitDots/Editor/Jobs/TriangleCountJob.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 36aeb96d38bf7ee4bb17e4fb5a1114e3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
