+++ UnitySDK/Packages/OpenBrushToolkitDots/Editor/Jobs/VertexMapJob.cs
