+++ UnitySDK/Packages/OpenBrushToolkitDots/Editor/Jobs/VertexMapJob.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5eae16851b7fec049ba22cee6149a581
+guid: 2c53daf1f9ccc6b4da3ea14d8fc22259
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
