+++ UnitySDK/Packages/OpenBrushToolkitDots/Editor/MathematicsUtils.cs
