+++ UnitySDK/Packages/OpenBrushToolkitDots/Editor/MathematicsUtils.cs.meta
@@ -1,3 +1,3 @@
 fileFormatVersion: 2
-guid: 2c1e255df445465d91a4d06af9b32d3a
+guid: b7b88490354493e42b044fd466015824
 timeCreated: 1734724080
