+++ UnitySDK/Packages/OpenBrushToolkitDots/Editor/SplitMeshByStroke.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7a9cc43d4fc33544488e220e3ee24b36
+guid: 3693b75eec3e7784cb4e988b28dcfecb
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
