+++ UnitySDK/Packages/OpenBrushToolkitDots/Editor/icosafoundation.openbrushtoolkitdots.editor.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: cab59376e5d607244b4ba972c04b73c3
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
