+++ UnitySDK/Packages/OpenBrushToolkitDots/package.json
@@ -1,0 +1,11 @@
+{
+    "name": "com.icosafoundation.openbrushtoolkitdots",
+    "version": "1.0.0",
+    "displayName": "Open Brush Toolkit DOTS",
+    "description": "This package contains Open Brush Toolkit features using Unity DOTS.",
+    "unity": "2019.4",
+    "unityRelease": "25f1",
+    "dependencies": {
+      "com.unity.burst": "1.4.11"
+    }
+  }
