+++ UnitySDK/Packages/OpenBrushToolkitDots/package.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 671302a52297c5b4c9816939cb1083d8
+PackageManifestImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
